Fix styling issues and moved PropertyContainsKeywordsPredicate to Property from Appointment.
