+++ src/main/java/seedu/address/logic/commands/FindPropertyCommand.java
@@ -1,10 +1,10 @@
 package seedu.address.logic.commands;
 
+import static java.util.Objects.requireNonNull;
+
 import seedu.address.commons.core.Messages;
 import seedu.address.model.Model;
-import seedu.address.model.appointment.PropertyContainsKeywordsPredicate;
-
-import static java.util.Objects.requireNonNull;
+import seedu.address.model.property.PropertyContainsKeywordsPredicate;
 
 /**
  * Finds and lists all persons in address book whose name contains any of the argument keywords.
