+++ src/main/java/seedu/address/logic/parser/FindPropertyCommandParser.java
@@ -1,13 +1,13 @@
 package seedu.address.logic.parser;
 
-import seedu.address.logic.commands.FindAppointmentCommand;
-import seedu.address.logic.commands.FindPropertyCommand;
-import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.appointment.PropertyContainsKeywordsPredicate;
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.Arrays;
 
-import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+import seedu.address.logic.commands.FindAppointmentCommand;
+import seedu.address.logic.commands.FindPropertyCommand;
+import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.property.PropertyContainsKeywordsPredicate;
 
 /**
  * Parses input arguments and creates a new FindCommand object
